flask-server/requirements.txt
flask==2.2.3
torch==2.0.0
torchvision==0.15.1
Pillow==9.5.0
flask-cors==3.0.10
numpy==1.24.3 